refactor: 상태 관리 전면 개선 - BaseModel 도입 및 코드 구조 정리
