+++ ai/node/apilist/map_node.py
@@ -1,7 +1,7 @@
 import logging
 from typing import Dict, Any
 
-from state.apilist.mapping_state import MappingState, update_map_success, update_map_failed
+from state.apilist.mapping_state_processor import MappingState, update_map_success, update_map_failed
 from service.apilist.map_agent import perform_semantic_mapping
 
 
@@ -13,13 +13,13 @@ def map_node(state: MappingState) -> Dict[str, Any]:
     logging.info("의미적 매핑 노드 시작")
     
     # retry_count 증가
-    retry_count = state.get("retry_count", 0) + 1
-    state["retry_count"] = retry_count
+    retry_count = (state.attempt_count or 0) + 1
+    state = state.model_copy(update={"attempt_count": retry_count})
     
     try:
         # 재생성 시에는 원본 데이터 사용, 처음에는 입력 데이터 사용
-        scenarios = state.get('scenarios', []) or state.get('original_scenarios', [])
-        api_lists = state.get('api_lists', []) or state.get('original_api_lists', [])
+        scenarios = state.scenarios or state.original_scenarios or []
+        api_lists = state.api_lists or state.original_api_lists or []
 
         if not scenarios or not api_lists:
             return update_map_failed(state, "시나리오 또는 API 목록이 없습니다")
