+++ ai/node/apilist/mapping_validation_node.py
@@ -1,6 +1,6 @@
 import logging
 from typing import Dict, Any
-from state.apilist.mapping_state import MappingState, update_mapping_validation_success, update_mapping_validation_failed
+from state.apilist.mapping_state_processor import MappingState, update_mapping_validation_success, update_mapping_validation_failed
 from service.apilist.mapping_validator import perform_mapping_validation
 
 
@@ -11,7 +11,7 @@ def mapping_validation_node(state: MappingState) -> Dict[str, Any]:
     """
     logging.info("매핑표 검증 노드 시작")
     try:
-        mapping_table = state.get("generated_mapping_table")
+        mapping_table = state.generated_mapping_table
         if not mapping_table:
             return update_mapping_validation_failed(state, "검증할 매핑표가 없습니다.")
